// apps/frontend/src/app/shared/components/overview/table-content-type.enum.ts
/*
 * Copyright Fraunhofer Institute for Material Flow and Logistics
 *
 * Licensed under the Apache License, Version 2.0 (the "License").
 * For details on the licensing terms, see the LICENSE file.
 * SPDX-License-Identifier: Apache-2.0
 */

export enum ContentType {
  COMPANIES = 'companies',
  PRODUCTS = 'products',
  PACKAGE = 'packaging',
  PRODUCT_GROUPS = 'product-groups',
  REPORTS = 'reports',
  BILLOFMATERIALS = 'billOfMaterials',
}